refac: Move store to i18n index
